+++ CHANGELOG.md
@@ -44,6 +44,9 @@
 * Avoid expensive set construction in Config constructor
   [#1289](https://github.com/bugsnag/bugsnag-android/pull/1289)
 
+* Replace calls to String.format() with concatenation
+  [#1293](https://github.com/bugsnag/bugsnag-android/pull/1293)
+
 ## 5.9.4 (2021-05-26)
 
 * Unity: add methods for setting autoNotify and autoDetectAnrs